@@ -9,22 +9,23 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "framer-motion": "^10.16.16",
-    "lucide-react": "^0.294.0",
+    "@tailwindcss/postcss": "^4.2.0",
+    "framer-motion": "^12.34.2",
+    "lucide-react": "^0.574.0",
     "next": "^16.1.6",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^19",
+    "react-dom": "^19",
     "react-icons": "^5.5.0"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "autoprefixer": "^10.0.1",
-    "eslint": "^4.1.1",
-    "eslint-config-next": "^0.2.4",
+    "@types/node": "^25",
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "autoprefixer": "^10.4.24",
+    "eslint": "^10.0.0",
+    "eslint-config-next": "^16.1.6",
     "postcss": "^8",
-    "tailwindcss": "^3.3.0",
+    "tailwindcss": "^4.2.0",
     "typescript": "^5"
   }
 }
